+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add InSpec test to verify JDK keystore password is not the infamous `changeit` #131
+
 ## 4.9.0 - 2019-08-16
 ### Added
 - Add AWS resources creation and deletion to integration testing